+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 CHANGELOG
 =========
 
+0.58.0
+------
+- Additional border and label for the list section
+    - Options
+        - `--list-border[=STYLE]`
+        - `--list-label=LABEL`
+        - `--list-label-pos=COL[:bottom]`
+    - Colors
+        - `list-fg`
+        - `list-bg`
+        - `list-border`
+        - `list-label`
+    - Actions
+        - `change-list-label`
+        - `transform-list-label`
+
 0.57.0
 ------
 - You can now resize the preview window by dragging the border